update aiter framework for a8w4 moe
